Add files via upload
edit_book.php
This is an admin-only page that allows a librarian to modify the details of an existing book. Its primary responsibilities are:

Pre-filled Form: It displays a form identical to the "Add Book" page, but with all the fields pre-populated with the book's existing data (Title, Author, Total Copies, etc.). This makes it easy for the librarian to see what they are editing.


Associated CSS:
admin_styles.css


Associated Backend:
backend/update_book.php: